backend/requirements.txt
fastapi
uvicorn[standard]
snowflake-connector-python[async]
python-dotenv
pydantic
aiohttp
